dslogic: Use actual captured samples in buffered mode
During buffered mode the internal memory of the DS
device may not be able to capture the amount of
samples requested. In that case, reading the samples
beyond max captured point will return all random data.

This commit reads the number of captures samples from the
trigger header. With that information the existing
limit_samples variable is overwritten.

The change only applies to buffered mode as in streaming
mode a overflow is automatically detected.