dslogic: Use queue for buffer deinterleaving in worker thread
Instead of only using one buffer for deinterleaving in worker thread, this
commit uses now a queue: For every finished USB transfer, the data
is put into a queue and a new transfer is immediately scheduled. This
completely decouples USB transfers from data processing activities
which could be quite time consuming, especially the deinterlaving
of data with high number of signal changes. The additional buffers
needed are dynamically allocated, i.e. as long as memory is available
it will allocate new buffers for the data processing or stops the acquisition
if no more memory is available. Once a buffer has finished prcessing,
the memory will be released immediately.